requirements.txt
eland
pandas
tensorflow
python-dotenv